requirements.txt
fastapi
uvicorn
streamlit
transformers
torch
datasets
pillow
python-multipart